fix: resolve mock.module conflicts between test files
Split llama wrapper tests into llama-wrapper.test.ts to avoid
Bun's global mock.module conflicts with embeddings.test.ts.
